requirements.txt
qrcode
Pillow
pyzbar
opencv-python